requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.11.12
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.8.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
python-toon==0.1.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
